+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 ## Next
 
+## 3.6.1 - 2024-08-30
+
 - fix: do not clear events when reset is called ([#170](https://github.com/PostHog/posthog-android/pull/170))
 - fix: reload feature flags as anon user after reset is called ([#170](https://github.com/PostHog/posthog-android/pull/170))